// src/__generated__/SimpleUserFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: SimpleUserFragment
// ====================================================

export interface SimpleUserFragment {
  __typename: "User";
  id: number;
  name: string;
  email: string;
  smallProfileImg: string | null;
}